+++ qqbot/apps.py
@@ -6,7 +6,7 @@ from django.apps import AppConfig
 class QqbotConfig(AppConfig):
     name = "qqbot"
     label = "qqbot"
-    verbose_name = "QQ 绑定"
+    verbose_name = "QQ binding / QQ 绑定"
     default_auto_field = "django.db.models.BigAutoField"
 
     def ready(self):